refactor polymer.py (second draft)
