+++ meeko/decoders.py
@@ -1,0 +1,270 @@
+import pathlib
+import json
+import logging
+import traceback
+from importlib.resources import files
+from os import linesep as eol
+from sys import exc_info
+from typing import Union
+from typing import Optional
+
+import rdkit.Chem
+from rdkit import Chem
+from rdkit.Chem import rdFMCS
+from rdkit.Chem import rdChemReactions
+from rdkit.Chem import rdMolInterchange
+from rdkit.Geometry import Point3D
+
+from .molsetup import RDKitMoleculeSetup
+from .molsetup import MoleculeSetupEncoder
+from .utils.jsonutils import rdkit_mol_from_json
+from .utils.rdkitutils import mini_periodic_table
+from .utils.rdkitutils import react_and_map
+from .utils.rdkitutils import AtomField
+from .utils.rdkitutils import build_one_rdkit_mol_per_altloc
+from .utils.rdkitutils import _aux_altloc_mol_build
+from .utils.rdkitutils import covalent_radius
+from .utils.pdbutils import PDBAtomInfo
+from .chemtempgen import export_chem_templates_to_json
+from .chemtempgen import build_noncovalent_CC
+from .chemtempgen import build_linked_CCs
+
+import numpy as np
+
+
+
+# region JSON Decoders
+
+
+def monomer_json_decoder(obj: dict):
+    """
+    Takes an object and attempts to decode it into a Monomer object.
+
+    Parameters
+    ----------
+    obj: Object
+        This can be any object, but it should be a dictionary generated by deserializing a JSON of a Monomer
+        object.
+
+    Returns
+    -------
+    If the input is a dictionary corresponding to a Monomer, will return a Monomer with data
+    populated from the dictionary. Otherwise, returns the input object.
+
+    """
+    # if the input object is not a dict, we know that it will not be parsable and is unlikely to be usable or
+    # safe data, so we should ignore it.
+    if type(obj) is not dict:
+        return obj
+
+    # check that all the keys we expect are in the object dictionary as a safety measure
+    expected_monomer_keys = {
+        "raw_rdkit_mol",
+        "rdkit_mol",
+        "mapidx_to_raw",
+        "residue_template_key",
+        "input_resname",
+        "atom_names",
+        "mapidx_from_raw",
+        "padded_mol",
+        "molsetup",
+        "is_flexres_atom",
+        "is_movable",
+        "molsetup_mapidx",
+    }
+
+    if set(obj.keys()) != expected_monomer_keys:
+        return obj
+    # Extracts init mols for Monomer:
+    raw_rdkit_mol = rdkit_mol_from_json(obj["raw_rdkit_mol"])
+    rdkit_mol = rdkit_mol_from_json(obj["rdkit_mol"])
+    if obj["mapidx_to_raw"] is None:
+        mapidx_to_raw = None
+    else:
+        mapidx_to_raw = {int(k): v for k, v in obj["mapidx_to_raw"].items()}
+
+    monomer = Monomer(raw_rdkit_mol, rdkit_mol, mapidx_to_raw)
+
+    # sets remaining properties from JSON
+    monomer.residue_template_key = obj["residue_template_key"]
+    monomer.input_resname = obj["input_resname"]
+    monomer.atom_names = obj["atom_names"]
+
+    if obj["mapidx_from_raw"] is None:
+        monomer.mapidx_from_raw = None
+    else:
+        monomer.mapidx_from_raw = {int(k): v for k, v in obj["mapidx_from_raw"].items()}
+
+    monomer.padded_mol = rdkit_mol_from_json(obj["padded_mol"])
+    monomer.molsetup = RDKitMoleculeSetup.from_json(obj["molsetup"])
+    if obj["molsetup_mapidx"] is None:
+        monomer.molsetup_mapidx = None
+    else:
+        monomer.molsetup_mapidx = {int(k): v for k, v in obj["molsetup_mapidx"].items()}
+
+    # boolean values
+    monomer.is_flexres_atom = obj["is_flexres_atom"]
+    monomer.is_movable = obj["is_movable"]
+
+    return monomer
+
+
+def residue_template_json_decoder(obj: dict):
+    """
+    Takes an object and attempts to decode it into a ResidueTemplate object.
+
+    Parameters
+    ----------
+    obj: Object
+        This can be any object, but it should be a dictionary constructed by deserializing the JSON representation of a
+        ResidueTemplate object.
+
+    Returns
+    -------
+    If the input is a dictionary corresponding to a ResidueTemplate, will return a ResidueTemplate with data populated
+    from the dictionary. Otherwise, returns the input object.
+    """
+    # if the input object is not a dict, we know that it will not be parsable and is unlikely to be usable or
+    # safe data, so we should ignore it.
+    if type(obj) is not dict:
+        return obj
+
+    # check that all the keys we expect are in the object dictionary as a safety measure
+    expected_residue_keys = {"mol", "link_labels", "atom_names"}
+    if set(obj.keys()) != expected_residue_keys:
+        return obj
+
+    # Converting ResidueTemplate init values that need conversion
+    deserialized_mol = rdkit_mol_from_json(obj["mol"])
+    # do not write canonical smiles to preserve original atom order
+    mol_smiles = rdkit.Chem.MolToSmiles(deserialized_mol, canonical=False)
+    link_labels = {int(k): v for k, v in obj["link_labels"].items()}
+
+    # Construct a ResidueTemplate object
+    residue_template = ResidueTemplate(mol_smiles, None, obj["atom_names"])
+    # Separately ensure that link_labels is restored to the value we expect it to be so there are not errors in
+    # the constructor
+    residue_template.link_labels = link_labels
+
+    return residue_template
+
+
+def residue_padder_json_decoder(obj: dict):
+    """
+    Takes an object and attempts to decode it into a ResiduePadder object.
+
+    Parameters
+    ----------
+    obj: Object
+        This can be any object, but it should be a dictionary constructed by deserializing the JSON representation of a
+        ResiduePadder object.
+
+    Returns
+    -------
+    If the input is a dictionary corresponding to a ResiduePadder, will return a ResiduePadder with data populated
+    from the dictionary. Otherwise, returns the input object.
+    """
+    # if the input object is not a dict, we know that it will not be parsable and is unlikely to be usable or
+    # safe data, so we should ignore it.
+    if type(obj) is not dict:
+        return obj
+
+    # check that all the keys we expect are in the object dictionary as a safety measure
+    expected_residue_keys = {
+        "rxn_smarts",
+        "adjacent_smarts",
+        "auto_blunt",
+    }
+    if set(obj.keys()) != expected_residue_keys:
+        return obj
+
+    # Constructs a ResiduePadder object and restores the expected attributes
+    # adjacent_smartsmol_mapidx is rebuilt by ResiduePadder init
+    residue_padder = ResiduePadder(obj["rxn_smarts"], obj["adjacent_smarts"], obj["auto_blunt"])
+
+    return residue_padder
+
+
+def residue_chem_templates_json_decoder(obj: dict):
+    """
+    Takes an object and attempts to decode it into a ResiduePadder object.
+
+    Parameters
+    ----------
+    obj: Object
+        This can be any object, but it should be a dictionary constructed by deserializing the JSON representation of a
+        ResiduePadder object.
+
+    Returns
+    -------
+    If the input is a dictionary corresponding to a ResiduePadder, will return a ResiduePadder with data populated
+    from the dictionary. Otherwise, returns the input object.
+    """
+    # if the input object is not a dict, we know that it will not be parsable and is unlikely to be usable or
+    # safe data, so we should ignore it.
+    if type(obj) is not dict:
+        return obj
+
+    # Check that all the keys we expect are in the object dictionary as a safety measure
+    expected_residue_keys = {
+        "residue_templates",
+        "ambiguous",
+        "padders",
+    }
+    if set(obj.keys()) != expected_residue_keys:
+        return obj
+
+    # Extracting the constructor args from the json representation and creating a ResidueChemTemplates instance
+    templates = {
+        k: residue_template_json_decoder(v) for k, v in obj["residue_templates"].items()
+    }
+    padders = {k: residue_padder_json_decoder(v) for k, v in obj["padders"].items()}
+
+    residue_chem_templates = ResidueChemTemplates(templates, padders, obj["ambiguous"])
+
+    return residue_chem_templates
+
+
+def polymer_json_decoder(obj: dict):
+    """
+    Takes an object and attempts to deserialize it into a Polymer object.
+
+    Parameters
+    ----------
+    obj: Object
+        This can be any object, but it should be a dictionary constructed by deserializing the JSON representation of a
+        Polymer object.
+
+    Returns
+    -------
+    If the input is a dictionary corresponding to a Polymer, will return a Polymer with data
+    populated from the dictionary. Otherwise, returns the input object.
+    """
+    # if the input object is not a dict, we know that it will not be parsable and is unlikely to be usable or
+    # safe data, so we should ignore it.
+    if type(obj) is not dict:
+        return obj
+
+    # Check that all the keys we expect are in the object dictionary as a safety measure
+    expected_json_keys = {
+        "residue_chem_templates",
+        "monomers",
+        "log",
+    }
+    if set(obj.keys()) != expected_json_keys:
+        return obj
+
+    # Deserializes ResidueChemTemplates from the dict to use as an input, then constructs a Polymer object
+    # and sets its values using deserialized JSON values.
+    residue_chem_templates = residue_chem_templates_json_decoder(
+        obj["residue_chem_templates"]
+    )
+
+    polymer = Polymer({}, {}, residue_chem_templates)
+    polymer.monomers = {
+        k: monomer_json_decoder(v) for k, v in obj["monomers"].items()
+    }
+    polymer.log = obj["log"]
+
+    return polymer
+# endregion
